Bug# 3111274 - Dump raw audio to file for automated audio quality analysis
This change is compiled for userdebug builds only.
The change enables raw audio capture of transcoded music.
The capture is written to file and used for audio quality analysis against the source audio.
This capture feature is OFF by default even for userdebug builds and requires setting a system property to enable it.

Change-Id: I6d895a2fe3ef9741fea490fe2513404a50c4d4b7