opensles: fix 86 compilation.
The libopensles Android.mk lists several C++ source files with
a .c extension. It also uses the '-x c++' compiler flag to force
their compilation as C++ sources.

For some reason, our x86 toolchain re-enables rtti in this case,
which leads to a library that cannot be linked properly. Fix the
issue by adding -fno-rtti when we're targetting the x86 CPU
architecture.

Change-Id: If00a86f93b492fc7abd97f92b5560138539bf1e0