Support for MediaRecorder stop exception in Camera
When the user of camera app presses "start" and "stop" very quickly with
the effects on, the MediaRecorder stop throws an exception since no
frames could be recorded before the user pressed stop. The
filterframework needs to take care of this issue separately as compared
to other RunTimeExceptions.

Related to bugs 5443546 and 5373197

Change-Id: I227fd1a81cbf9e48032e100111a896ffb691bf0c