Fixing a bug in gl_env release surface
While recording in camera with the effects on, there was a bug in releasing
a surface. The possible memory corruption was due to erasing
the surface-window pair in the map before destroying the surface and its associated
window handle.
Also, fixed a bit of nomenclature in the related functions.

Related to bug: 5373197

Change-Id: Ie861e5ce87cde8909c5680897cdfddafeba5e507