Fixing file sizelimit error in video capture intent with effects
There was an error in video capture intent with effects on
with regards to file sizelimit. This CL fixes that. Now the video gets
recorded without an error. The only remaining issue is that
the max file sizelimit is reached very quickly (~5s)  since
effects get recorded at 480p.

Related to bug: 5441217

Change-Id: I045ea1e2b6bb501c0f68954fdb28d135e9a4ed75