audio utils: fix resampler frame count calculation
The calculation of the number of input frames needed by
resampler_resample_from_provider() for a given number of output frames
is broken. The consequence is that we request too many frames when upsampling
causing unecessary buffering and delay and too few frames when downsampling causing
more calls to provider->get_next_buffer() than necessary.

Change-Id: I74b6a1708f347f45ca3c12312bf1b47320823742