Multi-Project Commit: Move of filterfw out of system/media (1 of 7)
This is part of the multi-project commit to move the filter-framework
from system/media/mca to frameworks/base/media/mca.

Note that the filter-framework will soon be replaced with a refactored
version currently under API review (also to go under frameworks/base).
This move is done now to unblock the PDK efforts.

Change-Id: Ia065ab69c2431886ae3a3db96290fa5e0f1e652d