audio_route: use mixer_ctl_get/set_array
Rather than make individual calls to tinyalsa for every
value in each mixer control, which is expensive for linked
controls, use the array get/set functions instead. This
reduces the number of system calls that tinyalsa has to
make. The majority of controls are linked and have 1 or 2
values so the CPU overhead is small.

Change-Id: If32cb8c639e113403641cab0d68b32799cebb059